backend/upload-photo/requirements.txt
Pillow==10.1.0
psycopg2-binary==2.9.9
